consul-kvsync: propagate deletes
When the source (remote) changes, it only returns the current values of
all the keys, but not the ones that were deleted. So we don't actually
see deletions, and so can't delete from the destination.

We now keep a watch on the destination (local) store for each path. When
it changes, we update an in-memory list of all the keys that exist
there.

Then, when we get changes from the source, we delete any keys from the
destination that aren't on the source.

Nice.